+++ bin/main.dart
@@ -10,10 +10,10 @@ void main(List<String> arguments) {
   print(
       'Gorkha Earthquake Details\nYear = ${gorkhaEarthQuake.year}\nMonth = ${gorkhaEarthQuake.month}\nDay = ${gorkhaEarthQuake.day}\nHour = ${gorkhaEarthQuake.hour}\nMinute = ${gorkhaEarthQuake.minute}');
   print('\nCurrent NepaliDateTime = ${NepaliDateTime.now()}');
-  print(gorkhaEarthQuake.mergeTime(10, 20, 30));
+  print('\nMerged NepaliDateTime = ${gorkhaEarthQuake.mergeTime(10, 20, 30)}');
 
   heading('Date Conversion');
-  var nt = NepaliDateTime.fromDateTime(DateTime(2019, 8, 03, 14, 30, 15));
+  var nt = DateTime(2019, 8, 03, 14, 30, 15).toNepaliDateTime();
   print('In BS = $nt');
   var dt = nt.toDateTime();
   print('In AD = $dt');